test(history): fixes tests and extends to 3 networks
Before this, tests had to be run manually with an environment variable `NETWORK`, and the test was run against the user submitted address.

Removes the need for `NETWORK` by adding 3 distinct test cases for each network.

Removes dependency on user submitted address, by hardcoding 3 address that definitely have a transaction history on the respective network.

This ensures the test asserts the desired behaviour and doesn't depend on user submitted values.