chore(transactions): simplifies and un-nests `getTxType`
The first if statement seems redundant, because if `isCIP64(tx)` is true, then
1. there cannot be a tx.gatewayFee,
2. there cannot be a tx.gatewayFeeRecipient
3. there cannot be a tx.gasPrice

It might have made sense to keep the code that deletes the fields in case `isCIP64(tx)` doesn't work as expected. But that adds more complexity than necessary in my opinion.

Un-nesting if, else if, else statement (personal preference).