test(transfers): changes ERC-20 transfer test
From cUSD to USDC transfer.
The test was and is still incomplete, because it was and still is only testing that an ERC-20 transfer can be made.

It doesn't assert that a USDC transfer with USDC as fee currency can be made.

Moved this test into the CIP-64 block for ease of reference.